feat: use Script API
