+++ gatsby-config.js
@@ -11,5 +11,6 @@ module.exports = {
         path: `${__dirname}/content/posts`
       }
     }
-  ]
+  ],
+  partytownProxiedURLs: ['https://utteranc.es/client.js']
 };
